moved original json schema generator to separate module
